Fix misc code to get rid of useless type conversions or suppress `REDUNDANT_CALL_OF_CONVERSION_METHOD`
It's necessary because the full Kotlin is compiled
by a bootstrap version and `-Werror` is enabled.

KT-77727

+++ analysis/low-level-api-fir/src/org/jetbrains/kotlin/analysis/low/level/api/fir/caches/FirThreadSafeCachesFactory.kt
@@ -61,7 +61,7 @@ internal class FirThreadSafeCachesFactory(private val project: Project) : FirCac
         }
 
         if (maximumSize != null) {
-            builder.maximumSize(maximumSize.toLong())
+            builder.maximumSize(maximumSize)
         }
 
         if (keyStrength == KeyReferenceStrength.WEAK) {
